diabetes-proyecto-3/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.40.45",
    "fastapi>=0.116.1",
    "gradio>=5.49.0",
    "mlflow==2.14.2",
    "pandas>=2.3.2",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn==1.7.1",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "dill"
]
